Rename original window to gui, implement Window struct
